Three tiny fixes (#3)
* Properly create asyncio event loop

In recent asyncio versions (in my case, 3.13), you need to create the
event loop yourself before calling `asyncio.get_event_loop()` or it will
emit a DeprecationWarning.

* Make main.py dump result in JSON

The README file implies main.py outputs JSON but it doesn't, and the
output could not be parsed properly with JSON tools, due to the type of
quotes used by print().

This commit makes main.py use the json module instead, so that the
output is valid JSON. While it is not strictly necessary we also use a
2-space indentation level for convenience.

* Fix bad details when there was nothing in one section

Previously when a section was empty we'd get a buggy "details" item:

    {
      "name": "RECYPARC",
      "start_date": "01-01-2025",
      "id": "RECYPARC",
      "details": [
        {
          "date": "",
          "detail": ""
        }
      ],
      "total": "1"
    }

This is due to the Intradel website adding an empty row in the table in
this case. As a valid row would need a date, we solve the problem by
filtering out rows without dates.